Refactor payments into charge list
